Clarify the TIMEOUT argument of ament_add_gtest
In my testing adding a TIMEOUT argument to ament_add_gtest actually adds that many seconds.  Also it must be specified as a floating point number for it to work.  I'm not sure why but the documentation could use some clarification.  Also, one thing to note is this timeout is per-test, not for the full gtest binary.  A gtest executable can contain many tests.